# docs/api_doc/swagger.json
{"openapi":"3.0.1","info":{"title":"'@project.name@'","description":"'@project.description@'","contact":{"name":"Joe Doe","email":"dumy@dumy"},"version":"'@project.version@'"},"servers":[{"url":"http://localhost:8080","description":"'@project.description@'"}],"paths":{"/reservations":{"put":{"tags":["reservation-controller"],"operationId":"checkInReservation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationDTO"}}}}}},"post":{"tags":["reservation-controller"],"operationId":"createReservation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationDTO"}}}}}}},"/flights":{"post":{"tags":["flight-controller"],"operationId":"postFlight","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FligthsResponseDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FligthsResponseDTO"}}}}}}},"/reservations/{id}":{"get":{"tags":["reservation-controller"],"operationId":"getReservation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationDTO"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/flights/{id}":{"get":{"tags":["flight-controller"],"operationId":"getFlights","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FligthsResponseDTO"}}}}}}}},"components":{"schemas":{"PassengerDTO":{"type":"object","properties":{"name":{"type":"string"},"seat":{"type":"string"}}},"ReservationDTO":{"type":"object","properties":{"id":{"type":"string"},"status":{"type":"string","enum":["READY-TO-CHECK-IN","CHECKED-IN","DUE"]},"flight":{"type":"string"},"passengers":{"type":"array","items":{"$ref":"#/components/schemas/PassengerDTO"}}}},"AirportDTO":{"type":"object","properties":{"name":{"type":"string"},"code":{"type":"string"},"location":{"type":"string"}}},"FligthsResponseDTO":{"type":"object","properties":{"id":{"type":"string"},"aircraft":{"type":"string"},"departure":{"type":"string","format":"date-time"},"airport":{"$ref":"#/components/schemas/AirportDTO"},"seat_map":{"type":"array","items":{"$ref":"#/components/schemas/SeatDTO"}}}},"SeatDTO":{"type":"object","properties":{"seat":{"type":"string"},"status":{"type":"string","enum":["reserved","available"]}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}