# src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/common/AirportDTO.java
package ar.edu.utn.frc.tup.lc.iv.dtos.common;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class AirportDTO {

    private String name;

    private String code;

    private String location;
}