# src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/common/PassengerDTO.java
package ar.edu.utn.frc.tup.lc.iv.dtos.common;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class PassengerDTO {

    private String name ;
    private String seat ;
}