# src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/common/SeatDTO.java
package ar.edu.utn.frc.tup.lc.iv.dtos.common;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class SeatDTO {

    private String seat;

    private SeatStatus status;
}